Fold the WOW field pavilion into WowProfileBody and retire the mobileProfile surface
